Added edit user password Hook
